src/fuzzy_tm_infer/tools/compression/requirements.txt
brotli
blosc2
cramjam
isal
lz4
matplotlib
python-snappy
zopfli
zstandard
